tracker_spend/backend_python/requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
requests==2.31.0
openpyxl==3.1.2
